docs: instructions on using external [N4S] rules
