+++ packages/n4s/config/rollup/rollup.config.js
@@ -1,4 +1,3 @@
 import enforce from './enforce';
-import enforceExtended from './enforceExtended';
 
-export default [].concat(enforce, enforceExtended);
+export default [].concat(enforce);
