+++ packages/n4s/docs/_sidebar.md
@@ -1,6 +1,6 @@
 - [Main](./)
 - [List of Enforce Rules](./rules)
-- [Business reated rules](./business_rules)
 - [Schema validations](./coumpound)
 - [Templates and Composites](./template)
 - [Custom Enforce Rules](./custom)
+- [Consuming external rules](./external)
