+++ packages/n4s/package.json
@@ -7,16 +7,6 @@
   "exports": {
     "./n4s.js": "./n4s.js",
     "./package.json": "./package.json",
-    "./enforceExtended.cjs.development.js": "./enforceExtended.cjs.development.js",
-    "./enforceExtended.cjs.production.js": "./enforceExtended.cjs.production.js",
-    "./enforceExtended.cjs.production.min.js": "./enforceExtended.cjs.production.min.js",
-    "./enforceExtended.js": "./enforceExtended.js",
-    "./enforceExtended.umd.development.js": "./enforceExtended.umd.development.js",
-    "./enforceExtended.umd.production.js": "./enforceExtended.umd.production.js",
-    "./enforceExtended.umd.production.min.js": "./enforceExtended.umd.production.min.js",
-    "./esm/enforceExtended.mjs.development.js": "./esm/enforceExtended.mjs.development.js",
-    "./esm/enforceExtended.mjs.production.js": "./esm/enforceExtended.mjs.production.js",
-    "./esm/enforceExtended.mjs.production.min.js": "./esm/enforceExtended.mjs.production.min.js",
     "./esm/n4s.mjs.development.js": "./esm/n4s.mjs.development.js",
     "./esm/n4s.mjs.production.js": "./esm/n4s.mjs.production.js",
     "./esm/n4s.mjs.production.min.js": "./esm/n4s.mjs.production.min.js",
