+++ packages/vest/docs/enforce.md
@@ -15,3 +15,4 @@ test('username', 'Must be at least three characters long', () => {
   - [Creating Custom Rules](./n4s/custom)
   - [Shape and schema validation](./n4s/compound)
   - [Enforce templates](./n4s/template)
+  - [Consuming external rules](./n4s/external)
