--- packages/vest/src/utilities/enforceExtended@n4s.js
@@ -1,3 +1,0 @@
-import enforceExtended from 'enforceExtended';
-
-export default enforceExtended;
